split codegen into own file
